GitHub Actions, Python 3.9+ Support, Jupyterlab Extensions (#991)
* GitHub Actions updates, Conda environment file

* readme package info table

* readme fix

* unpinned ipywidgets

* testing windows path fix

* set cell_cluster_col type to np.int64

* set cell_cluster_col type to np.int64 v2

* dtype for pandas merge_asof explictly converted

* dtype for pandas merge_asof convert to int

* convert to int

* huggingface cache

* shutil fix

* change type to int

* changed a type to int

* example dataset cache download script

* moved getting the dataset from test to ci workflow file

* pycodestyle

* cache in ci fix

* python-version fix

* correct data path

* caching logic updates

* save / restore cache

* set cache dir to ./data/cache

* removed exampled dataset download in `fiber_segmentation_test`

* windows caching test

* remove misc notebook metadata

* import json fix

* test dtype change

* example dataset optimizations

* type hinting

* type hinting 2

* type hinting 3

* pycodestyle

* dimensionality_reduction_test speed up

* parallel tests

* removed parallel tests, unfortunatly it hangs indefinity on windows

* removed xdist decorators

* removed xdist decorators 2

* proper path for cache_dir on windows

* windows paths

* no posix() path

* pytest-xdist test 2

* added docker badge

* added jupyterlab extensions as a set of optional dependencies

* remove unused imports

* remove unused imports

* sort imports

* environment.yml updates

* dockerfile updates

* removed duplicate markdown section

* updated cibuildwheel

* type def updates

* fixed pandas dtype for `calculate_mean_distance_to_all_cell_types`

* type fix

* make sure args of watershed are np.int32

* set pixie types to int for som cluster

* removed xdist, when the notebooks test in parallel, it copies data for each -> no more space on the runner

+++ src/ark/analysis/cell_neighborhood_stats.py
@@ -1,11 +1,12 @@
 import os
+from functools import reduce
+
 import numpy as np
 import pandas as pd
 import xarray as xr
-from functools import reduce
+from alpineer import io_utils, misc_utils
 
 import ark.settings as settings
-from alpineer import misc_utils, io_utils
 
 
 def shannon_diversity(proportions):
@@ -176,7 +177,7 @@ def calculate_mean_distance_to_all_cell_types(
     all_clusters = np.unique(cell_table[cell_type_col])
 
     # call calculate_mean_distance_to_cell_type for all cell clusters
-    avg_dists = pd.DataFrame(index=cell_table.index.values, columns=all_clusters)
+    avg_dists = pd.DataFrame(index=cell_table.index.values, columns=all_clusters, dtype=np.float64)
     for cell_cluster in all_clusters:
         avg_dists.loc[:, cell_cluster] = calculate_mean_distance_to_cell_type(
             cell_table, dist_xr, cell_cluster, k, cell_type_col, cell_label_col)
